# server/openapi.yaml
openapi: 3.0.3
info:
  title: Audiophile E-Commerce API
  version: 1.0.0
  description: This is an API for an Audio E-Commerce Site.
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        name:
          type: string
          description: Error name
        status:
          type: number
          description: HTTP status code
...
    Cart:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        active:
          type: boolean
...
paths:
...
  /cart/{userId}:
    post:
      summary: Create a new cart for a user
      description: Creates a new cart for a user. Does not create a new cart if there's an existing active cart.
      operationId: createCart
      parameters:
        - name: userId
          in: path
          description: Unique ID of the user
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Cart created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: User already has an active cart or invalid user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Get active cart for a specific user
      operationId: getActiveCart
      parameters:
        - name: userId
          in: path
          description: Unique ID of the user
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "404":
          description: User does not have an active cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/{userId}/add/{productId}:
    post:
      summary: Add a product to a cart
      description: Adds a product to a cart. The quantity of the product to be added is specified in the request body.
      operationId: addProductToCart
      parameters:
        - name: userId
          in: path
          description: Unique ID of the user
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: Unique ID of the product
          required: true
          schema:
            type: integer
      requestBody:
        description: Quantity of the product to be added to the cart
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 100
      responses:
        "200":
          description: Product added to cart successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid user ID, invalid product ID, or invalid quantity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User does not have an active cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/{userId}/remove/{productId}:
    delete:
      summary: Remove a product from the active cart of a user
      operationId: removeProductFromCart
      parameters:
        - name: userId
          in: path
          description: Unique ID of the user
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: Unique ID of the product
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product removed from cart successfully
        "400":
          description: Invalid user ID or invalid product ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User does not have an active cart or product is not in the cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/{userId}/checkout:
    post:
      summary: Checkout the active cart of a user
      description: Checks out the active cart of a user. The active cart is then marked as inactive.
      operationId: checkoutCart
      parameters:
        - name: userId
          in: path
          description: Unique ID of the user
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cart checked out successfully
        "400":
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User does not have an active cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...